+++ src/stdlib/general.rs
@@ -200,9 +200,6 @@ pub(crate) fn try_(vm: &mut VM, (context, handler): (RantValue, Option<RantFunct
     RantValue::Function(func) => {
       vm.call_func(func, vec![], PrintFlag::Sink, false)?;
     },
-    RantValue::Block(block) => {
-      vm.pre_push_block(&block, PrintFlag::Sink)?;
-    }
     other => runtime_error!(RuntimeErrorType::ArgumentError, "try: cannot protect '{}' value; only functions and blocks can be protected", other.get_type())
   }
   Ok(())